+++ src/main/java/spark/servlets/ByCountryServlet.java
@@ -1,19 +1,22 @@
 package spark.servlets;
 
-import spark.RDDCustomOperations;
-import spark.CustomLists.CustomListString;
+import spark.customlists.CustomListString;
+import spark.database.DatabaseOperations;
+import spark.rddoperations.RDDCustomOperations;
+import spark.rddoperations.SparkOperations;
 
 import java.io.IOException;
+import java.sql.SQLException;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.spark.api.java.JavaSparkContext;
-
 import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
+import org.apache.spark.api.java.JavaSparkContext;
+
 
 public class ByCountryServlet extends HttpServlet {
 
@@ -27,50 +30,43 @@ public class ByCountryServlet extends HttpServlet {
         this.sparkContext = context;
         this.inputFile = rdd;
     }
-    
+
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException 
-    {
-        resp.getWriter().println("Here is the sightingsByCountry page");
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        resp.getWriter().println("This is the ByCountryServlet");
         resp.getWriter().println("List of Commands: ");
         resp.getWriter().println("?inputType=inCountry&country=(yourcountryhere)    returns number of sightings in a country");
-        resp.getWriter().println("?inputType=numberOfCases&country=(yourcountryhere)     returns the number of sightings in a country");
         resp.getWriter().println("?inputType=byCountry         returns the number of sightings in each country");
         resp.getWriter().println();
-        resp.getWriter().println("List of mappings: time,state,project");
+        resp.getWriter().println("Valid mappings: time,country,state,shape,duration");
         resp.getWriter().println();
         resp.getWriter().println();
 
-        if(req.getParameter("inputType").equals("inCountry"))
-        {
-            resp.getWriter().println("Number of Cases in " + req.getParameter("country"));
-            resp.getWriter().println(numberOfCasesInCountry(req.getParameter("country")));
-            resp.getWriter().println();
+        if (req.getParameter("inputType").equals("inCountry")) {
             resp.getWriter().println("Number of sightings in " + req.getParameter("country"));
-            resp.getWriter().println(new CustomListString(sightingsInCountry(req.getParameter("country")).sortByKey().collect()));
-        }
-        else if(req.getParameter("inputType").equals("byCountry"))
-        {
+            try {
+                resp.getWriter().println(DatabaseOperations.readFromDatabase("byCountryTable", req.getParameter("country")));
+                resp.getWriter().println(new CustomListString(sightingsInCountry(req.getParameter("country")).sortByKey().collect()));
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        } else if (req.getParameter("inputType").equals("byCountry")) {
             resp.getWriter().println("Sightings by country:");
-            resp.getWriter().println(new CustomListString(sightingsByCountry().sortByKey().collect()));
+            try {
+                resp.getWriter().println(DatabaseOperations.printDatabase("byCountryTable","string"));
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+            
         }
         else
         {
             resp.getWriter().println("Invalid input.");
         }
-    }
-
-    public JavaPairRDD<String,Integer> sightingsByCountry()
-    {
-        JavaRDD<String> rdd2 =  RDDCustomOperations.rddStripToColumn(inputFile, 3);
-        JavaPairRDD<String,Integer> rdd3 = RDDCustomOperations.rddCounterString(rdd2);
- 
-        return rdd3;
-    }
-
-    public JavaPairRDD<String,Integer> sightingsInCountry(String country)
-    {
-        JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(inputFile,country.toLowerCase(),3);
+    } 
+    
+    public JavaPairRDD<String, Integer> sightingsInCountry(String country) {
+        JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(inputFile, country.toLowerCase(), 3);
 
         JavaRDD<String> rdd3 = rdd2.map(x ->
         {
@@ -79,14 +75,5 @@ public class ByCountryServlet extends HttpServlet {
         });
 
         return RDDCustomOperations.rddCounterString(rdd3);
-    }
-
-    public int numberOfCasesInCountry(String country)
-    {
-        JavaRDD<String> rdd2 = RDDCustomOperations.rddFiltering(inputFile,country.toLowerCase(),3);
-        JavaRDD<String> rdd3 = RDDCustomOperations.rddStripToColumn(rdd2,3);
-        JavaPairRDD<String,Integer> rdd4 = RDDCustomOperations.rddCounterString(rdd3);
-
-        return rdd4.lookup(country.toLowerCase()).get(0);
     }
 }